// backend/src/orders/order-status.enum.ts
export enum OrderStatus {
  Pending = 'PENDING',
  PaymentInProgress = 'PAYMENT_IN_PROGRESS',
  Paid = 'PAID',
  EnrollmentInProgress = 'ENROLLMENT_IN_PROGRESS',
  Completed = 'COMPLETED',
  PaymentFailed = 'PAYMENT_FAILED',
  RefundInProgress = 'REFUND_IN_PROGRESS',
  Refunded = 'REFUNDED',
  Failed = 'FAILED',
}